refactor: concern-split three near-cap AI lib modules (#4310 Slice B)
Extract economy production-boost resolution, ExpandMilitaryPlan value type
and war-count helpers, and colonial acquisition types for durable headroom
under the 400-line source ceiling. No AI semantics change.

Refs #4310
